+++ src/Modules/Ignis/Resources/CampaignResource.php
@@ -22,4 +22,14 @@ final class CampaignResource
             'groups/'.$this->groupId.'/campaigns/'.$this->campaignId,
         );
     }
+
+    public function playlogs(): CampaignPlaylogsResource
+    {
+        return new CampaignPlaylogsResource($this->client, $this->groupId, $this->campaignId);
+    }
+
+    public function contents(): CampaignContentsResource
+    {
+        return new CampaignContentsResource($this->client, $this->groupId, $this->campaignId);
+    }
 }